rename file to manager
