+++ src/packages/core/modal/index.ts
@@ -1,4 +1,4 @@
-export * from './modal.context.js';
+export * from './modal-manager.context.js';
 export * from './modal-handler.js';
 export * from './modal-route-registration.js';
 export * from './modal-route-registration.controller.js';
