+++ src/packages/core/modal/modal-handler.ts
@@ -1,4 +1,4 @@
-import { UmbModalConfig, UmbModalType } from './modal.context.js';
+import { UmbModalConfig, UmbModalType } from './modal-manager.context.js';
 import { UmbModalToken } from './token/modal-token.js';
 import type { IRouterSlot } from '@umbraco-cms/backoffice/external/router-slot';
 import type {
